Feat[UI]: moved Container and Logo components to UI directory
